requirements.txt
Django==4.2.3
sentry-sdk==1.29.0
pillow==10.0.0
coverage==7.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2 
pyjwt==2.8.0
drf-spectacular==0.26.4
django-filter==23.2